+++ networks/powerlaw/actual_burner.F90
@@ -31,6 +31,8 @@ contains
     type (burn_t),    intent(inout) :: state_out
     real(rt)        , intent(in   ) :: dt, time
 
+    !$gpu
+
     call integrator(state_in, state_out, dt, time)
 
   end subroutine actual_burner