+++ screening/screen_data.cpp
@@ -2,6 +2,8 @@
 
 using namespace amrex;
 
-#if NSCREEN > 0
-AMREX_GPU_MANAGED amrex::GpuArray<screen_factors_t, NSCREEN> scn_facs;
+namespace scrn {
+#if NUMSCREEN > 0
+    AMREX_GPU_MANAGED amrex::GpuArray<screen_factors_t, NSCREEN> scn_facs;
 #endif
+};