+++ unit_test/test_react/main.cpp
@@ -12,15 +12,21 @@ using namespace amrex;
 
 #include "test_react.H"
 #include "test_react_F.H"
+#include "extern_parameters.H"
+#include "eos.H"
+#include "network.H"
+#ifdef CXX_REACTIONS
+#include "react_zones.H"
+#endif
 #include "AMReX_buildInfo.H"
 
 
 int main (int argc, char* argv[])
 {
     amrex::Initialize(argc, argv);
-
-    main_main();
-
+    {
+        main_main();
+    }
     amrex::Finalize();
     return 0;
 }
@@ -36,6 +42,10 @@ void main_main ()
     std::string prefix = "plt";
 
     IntVect tile_size(1024, 8, 8);
+
+#ifdef CXX_REACTIONS
+    int do_cxx = 0;
+#endif
 
     // inputs parameters
     {
@@ -54,6 +64,10 @@ void main_main ()
         pp.query("max_grid_size", max_grid_size);
 
         pp.query("prefix", prefix);
+
+#ifdef CXX_REACTIONS
+        pp.query("do_cxx", do_cxx);
+#endif
 
     }
 
@@ -108,6 +122,17 @@ void main_main ()
 
     init_unit_test(probin_file_name.dataPtr(), &probin_file_length);
 
+    // Copy extern parameters from Fortran to C++
+    init_extern_parameters();
+
+    // C++ EOS initialization (must be done after Fortran eos_init and init_extern_parameters)
+    eos_init();
+
+#ifdef CXX_REACTIONS
+    // C++ Network, RHS, screening, rates initialization
+    network_init();
+#endif
+
     // Ncomp = number of components for each array
     int Ncomp = -1;
     init_variables();
@@ -152,6 +177,8 @@ void main_main ()
     // What time is it now?  We'll use this to compute total react time.
     Real strt_time = ParallelDescriptor::second();
 
+    int num_failed = 0;
+
     // Do the reactions
 #ifdef _OPENMP
 #pragma omp parallel
@@ -160,14 +187,43 @@ void main_main ()
     {
         const Box& bx = mfi.tilebox();
 
+#ifdef CXX_REACTIONS
+        if (do_cxx) {
+
+            auto s = state.array(mfi);
+            auto n_rhs = integrator_n_rhs.array(mfi);
+
+            int* num_failed_d = AMREX_MFITER_REDUCE_SUM(&num_failed);
+
+            AMREX_PARALLEL_FOR_3D(bx, i, j, k,
+            {
+                bool success = do_react(i, j, k, s, n_rhs);
+
+                if (!success) {
+                    Gpu::Atomic::Add(num_failed_d, 1);
+                }
+            });
+
+        }
+        else {
+#endif
+
 #pragma gpu
-        do_react(AMREX_INT_ANYD(bx.loVect()), AMREX_INT_ANYD(bx.hiVect()),
-                 BL_TO_FORTRAN_ANYD(state[mfi]),
-		 BL_TO_FORTRAN_ANYD(integrator_n_rhs[mfi]));
+            do_react(AMREX_INT_ANYD(bx.loVect()), AMREX_INT_ANYD(bx.hiVect()),
+                     BL_TO_FORTRAN_ANYD(state[mfi]),
+                     BL_TO_FORTRAN_ANYD(integrator_n_rhs[mfi]));
+
+#ifdef CXX_REACTIONS
+        }
+#endif
 
 	if (print_every_nrhs != 0)
 	  print_nrhs(AMREX_ARLIM_ANYD(bx.loVect()), AMREX_ARLIM_ANYD(bx.hiVect()),
 		     BL_TO_FORTRAN_ANYD(integrator_n_rhs[mfi]));
+    }
+
+    if (num_failed > 0) {
+        amrex::Abort("Integration failed");
     }
 
     // Call the timer again and compute the maximum difference between
@@ -196,10 +252,16 @@ void main_main ()
     std::string name = "test_react.";
     std::string integrator = buildInfoGetModuleVal(int_idx);
 
+#ifdef CXX_REACTIONS
+    std::string language = do_cxx == 1 ? ".cxx" : "";
+#else
+    std::string language = "";
+#endif
+
     // Write a plotfile
     int n = 0;
 
-    WriteSingleLevelPlotfile(prefix + name + integrator, state, varnames, geom, time, 0);
+    WriteSingleLevelPlotfile(prefix + name + integrator + language, state, varnames, geom, time, 0);
 
     // Tell the I/O Processor to write out the "run time"
     amrex::Print() << "Run time = " << stop_time << std::endl;